flowedit: Extract library_mgmt module from main.rs
Move library catalog management functions into a dedicated library_mgmt.rs
module: load_library_catalogs, add_library_function, resolve_node_source.
Associated tests moved to library_mgmt::test and flow_io::test modules.
